+++ test/endToEnd.js
@@ -32,7 +32,7 @@ describe('bml', function() {
   it('can process recursive rule choices', function() {
     let testString =
         `mode test {
-            (recurse!) as (just kidding) 50, (outer {(inner 1), (inner 2)}) 50
+            (recurse!) as {(just kidding) 50, (outer {(inner 1), (inner 2)}) 50}
         }
         {use test}
         recurse!
@@ -49,7 +49,7 @@ describe('bml', function() {
   it('respects the active mode on recursively rendered text', function() {
     let testString =
         `mode test {
-            (foo) as (bar) 100
+            (foo) as {(bar) 100}
         }
         {use test}
         {(foo)}